fix(wallet-utils): move amounts utils from accountsUtils to amountsUtils to prevent circular esm
