+++ scripts/circular-dependencies/madge-snapshot.txt
@@ -80,7 +80,5 @@ suite-common/wallet-core/src/transactions/transactionsSelectors.ts > suite-commo
 suite-common/wallet-core/src/transactions/transactionsSelectors.ts > suite-common/wallet-core/src/transactions/transactionsReducer.ts > suite-common/wallet-core/src/transactions/transactionsThunks.ts > suite-common/wallet-core/src/send/sendFormEthereumThunks.ts
 suite-common/wallet-types/src/index.ts > suite-common/wallet-types/src/solanaStaking.ts > suite-common/staking-solana/src/index.ts > suite-common/staking-solana/src/utils/index.ts > suite-common/staking-solana/src/utils/stakingUtils.ts
 suite-common/wallet-types/src/sendForm.ts > suite-common/wallet-types/src/transaction.ts
-suite-common/wallet-utils/src/accountUtils.ts > suite-common/wallet-utils/src/stakingUtils.ts > suite-common/wallet-utils/src/solanaStakingUtils.ts
-suite-common/wallet-utils/src/accountUtils.ts > suite-common/wallet-utils/src/transactionUtils.ts
 suite-common/wallet-utils/src/index.ts > suite-common/wallet-utils/src/accountUtils.ts > suite-common/suite-types/src/index.ts > suite-common/suite-types/src/modal.ts > suite-common/wallet-types/src/index.ts > suite-common/wallet-types/src/solanaStaking.ts > suite-common/staking-solana/src/index.ts > suite-common/staking-solana/src/utils/index.ts > suite-common/staking-solana/src/utils/stakingUtils.ts
 suite-native/device/src/index.ts > suite-native/device/src/hooks/useDetectDeviceError.tsx > suite-native/device/src/selectors.ts > suite-native/firmware/src/index.ts > suite-native/firmware/src/components/FirmwareInstallationScreenContent.tsx
