+++ suite-native/assets/tsconfig.json
@@ -19,10 +19,10 @@
             "path": "../../suite-common/wallet-core"
         },
         {
-            "path": "../../suite-common/wallet-utils"
+            "path": "../../suite-common/wallet-types"
         },
         {
-            "path": "../../suite-common/wallet-types"
+            "path": "../../suite-common/wallet-utils"
         },
         { "path": "../accounts" },
         { "path": "../atoms" },
