+++ suite-native/device-manager/tsconfig.json
@@ -12,10 +12,10 @@
             "path": "../../suite-common/wallet-core"
         },
         {
-            "path": "../../suite-common/wallet-utils"
+            "path": "../../suite-common/wallet-types"
         },
         {
-            "path": "../../suite-common/wallet-types"
+            "path": "../../suite-common/wallet-utils"
         },
         { "path": "../analytics" },
         { "path": "../atoms" },
